+++ libraries/VarAPI/include/var/StringView.hpp
@@ -27,6 +27,10 @@ class PathString;
 class NameString;
 class GeneralString;
 
+class StringView;
+using StringViewList = Vector<StringView>;
+
+
 /*! \brief StringView (wrapper for std::string_view)
  *
  *
@@ -38,7 +42,6 @@ class GeneralString;
  * value rather than reference.
  *
  */
-
 class StringView {
 public:
   constexpr static size_t npos = std::string_view::npos;
@@ -51,8 +54,8 @@ public:
    * The default for `std::string_view` is to create
    * a null string i.e `data() == nullptr`.
    */
-  StringView() : m_string_view{""}{}
-  StringView(const char *value) : m_string_view(value) {}
+  constexpr StringView() : m_string_view{""}{}
+  constexpr StringView(const char *value) : m_string_view(value) {}
   StringView(const String &value);
   StringView(const IdString &value);
   StringView(const KeyString &value);
@@ -60,10 +63,14 @@ public:
   StringView(const PathString &value);
   StringView(const NameString &value);
   StringView(const GeneralString &value);
-  explicit StringView(const std::string_view string_view)
+  explicit constexpr StringView(const std::string_view string_view)
     : m_string_view(string_view) {}
 
   StringView(const char *value, size_t length) : m_string_view{value, length} {}
+
+  static constexpr StringView whitespace(){
+    return StringView{"\t\n\v\f\r "};
+  }
 
   /*! \details This allows you to quickly check for an empty string.
    *
@@ -106,6 +113,18 @@ public:
 
   API_NO_DISCARD bool is_empty() const { return m_string_view.empty(); }
 
+  StringView& strip_leading_containing(StringView value);
+
+  StringView &strip_trailing_containing(StringView value);
+
+  StringView &strip_leading_whitespace(){
+    return strip_leading_containing(whitespace());
+  }
+
+  StringView &strip_trailing_whitespace(){
+    return strip_trailing_containing(whitespace());
+  }
+
   StringView &pop_front(size_t length = 1) {
     m_string_view.remove_prefix(length);
     return *this;
@@ -136,7 +155,7 @@ public:
   API_NO_DISCARD StringView get_substring_at_position(size_t position) const;
   API_NO_DISCARD StringView get_substring_with_length(size_t length) const;
 
-  API_NO_DISCARD var::Vector<StringView> split(StringView delimeters) const;
+  API_NO_DISCARD StringViewList split(StringView delimeters) const;
 
   using iterator = typename std::string_view::iterator;
   using const_iterator = typename std::string_view::const_iterator;
@@ -208,12 +227,12 @@ public:
     return m_string_view.rfind(a, position);
   }
 
-  API_NO_DISCARD size_t find_last_of(StringView a, size_t position = 0) const {
+  API_NO_DISCARD size_t find_last_of(StringView a, size_t position = npos) const {
     return m_string_view.find_last_of(a.m_string_view, position);
   }
 
   API_NO_DISCARD size_t
-  find_last_not_of(StringView a, size_t position = 0) const {
+  find_last_not_of(StringView a, size_t position = npos) const {
     return m_string_view.find_last_not_of(a.m_string_view, position);
   }
 
@@ -226,6 +245,10 @@ public:
   }
 
   API_NO_DISCARD bool contains(const StringView a) const {
+    return find(a) != npos;
+  }
+
+  API_NO_DISCARD bool contains(char a) const {
     return find(a) != npos;
   }
 
@@ -293,7 +316,6 @@ private:
 
 inline bool operator==(const char *lhs, StringView rhs) { return rhs == lhs; }
 
-using StringViewList = Vector<StringView>;
 
 inline StringView
 operator"" _string_view(const char * value) {